+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
  * improve frontend page templates render
  * add trasnlatable fields
  * add Page template types management & make frontend templates render according to it
+ * add NewsletterPost admin
 
 ##### Version 5.03.00 (2021-08-04)
  * add sonata admin dependency